% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UEFT.R
\name{UEFT}
\alias{UEFT}
\title{Uniform Exact Functional Test on Two Discrete Random Variables}
\usage{
UEFT(input, correct, log.p)
}
\arguments{
\item{input}{A matrix of nonnegative integers representing a contingency table.
Column is the casual and row is the effect.}

\item{correct}{Logical; if implement the continuity correction.
The description is at details. The default is TRUE.}

\item{log.p}{Logical; if TRUE, the p-value is given as log(p). The default is FALSE. The default is FALSE.}
}
\value{
The exact p-value of the test.
}
\description{
Perform the uniform exact functional test on a contingency table to determine if the
column variable is a function of the row variable.
}
\details{
The uniform idea was implementated using uniform marginal distribution of a square table as null hypothesis.
}
\note{
The functions provide a direct entry into the C++ implementations of the exact functional test.
}
\examples{
 # Initial a table
 x = matrix(c(0,5,10,0,0,5), ncol=3)
 # With continuity correction
 UEFT(x)
 # Without continuity correction
 UEFT(x, FALSE)
}
\author{
Yiyi Li, Joe Song
}
\keyword{exact}
\keyword{functional}
\keyword{nonparametric}
\keyword{test}
\keyword{uniform}
